정온 대댓글 작업끝

+++ www/src/main/java/com/booktopia/www/service/CommentServiceImpl.java
@@ -1,14 +1,17 @@
 package com.booktopia.www.service;
 
 import com.booktopia.www.domain.CommentVO;
+import com.booktopia.www.domain.DTO.CommentDTO;
 import com.booktopia.www.domain.PagingVO;
 import com.booktopia.www.domain.RecommentVO;
 import com.booktopia.www.handler.PagingHandler;
 import com.booktopia.www.repository.CommentMapper;
+import com.booktopia.www.repository.ReCommentMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -17,6 +20,7 @@ import java.util.List;
 public class CommentServiceImpl implements CommentService{
 
     private final CommentMapper commentMapper;
+    private final ReCommentMapper reCommentMapper;
 
     @Override
     public int post(CommentVO cvo) {
@@ -25,10 +29,25 @@ public class CommentServiceImpl implements CommentService{
 
     @Override
     public PagingHandler getCommentList(int bno, PagingVO pgvo) {
-        List<CommentVO> clist = commentMapper.getCommentList(bno,pgvo);
         //totalCount
         int totalCount = commentMapper.getSelectOneComment(bno);
-        PagingHandler ph = new PagingHandler(pgvo, totalCount, clist);
+        List<CommentVO> clist = commentMapper.getCommentList(bno, pgvo);
+        List<CommentDTO> cdtoList = new ArrayList<>();
+        if(clist.size()!=0){
+            for(CommentVO cvo : clist){
+                CommentDTO cdto = new CommentDTO();
+                long cno = cvo.getCno();
+                List<RecommentVO> rlist = reCommentMapper.getReCommentList(cno, pgvo);
+                cdto.setCvo(cvo);
+                cdto.setRclist(rlist);
+                cdtoList.add(cdto);
+                log.info("cno>>>>{}",cno);
+                log.info("rlist>>>>{}",rlist);
+            }
+        }
+        log.info("clist>>>>{}",clist);
+        log.info("cdtoList>>>>{}",cdtoList);
+        PagingHandler ph = new PagingHandler(pgvo, totalCount, (List<CommentDTO>) cdtoList);
         return ph;
     }
 
@@ -39,7 +58,11 @@ public class CommentServiceImpl implements CommentService{
 
     @Override
     public int deleteComment(long cno) {
-        return commentMapper.deleteComment(cno);
+        int isOk = reCommentMapper.deleteReComment(cno);
+        if(isOk>0){
+                commentMapper.deleteComment(cno);
+        }
+        return isOk;
     }
 
     @Override
